Get rid of allowed_to scope since it is not needed and queries are too different
